Added Lift and Collector
Added a basic lift with one motor and a collector with two CR servos(assuming they are vex) so far. Also added in some auto methods for our drivetrain. Will plan to work on them some more and fine tune them once we get the drivetrain. The goal is to work on gyro turning for next week